Sync 3: main ea6cb01 -> fe8bdd5 (15 commits). Clean automatic merge, no
conflicts (.gitignore and daemon/main.c auto-merged). What lands here that the
ARM port has to answer for, in descending order of blast radius:

- bitcoin_interp.asm (+326, the Core script differential, audit 6.9 + the
  verify-differential extension): MINIMALDATA now enforces minimal scriptnum
  ENCODING on operand reads (scriptnum_decode takes bit 8 of maxsize, SNUM_MAX
  at 12 read sites, PICK/ROLL/CLTV/CSV/WITHIN/CHECKSIGADD/both multisig counts
  report snum_overflow -> SCRIPTNUM); MINIMALIF returns SCRIPT_ERR_MINIMALIF;
  new interp_pubkey_encoding_ok for STRICTENC; LOW_S (<= n/2) and STRICTENC
  hashtype in interp_sig_encoding_ok (-1/-2 mapped to SIG_HIGH_S/SIG_HASHTYPE);
  BIP146 NULLFAIL; CONST_SCRIPTCODE/SIG_FINDANDDELETE (sentinel -5); CHECKMULTI-
  SIG's NULLDUMMY deferred behind the sig loop; DISCOURAGE_OP_SUCCESS moved to
  bit 19.
- CHECKSEQUENCEVERIFY consensus false-accept: `test rax, 0x80000000` assembled
  as a 5-byte test against 0xFFFFFFFF80000000, so a bits-32-39 operand with
  bit 31 clear was treated as disabled -> NOP -> accepted where Core enforces
  (tests/test_csv_disable_flag pins it).
- CompactSize bounds: varint_put (serve, cmpct) and write_varint (sighash)
  compared against 0xffffffff, which sign-extends to -1, so the 9-byte form was
  unreachable; bip152_shortid's `and rax, 0xffffffffffff` assembled to `and
  rax, -1` (no-op).
- scriptcodec: scriptnum_decode's minimal-encoding arm, and check_minimal_push
  where the OP_1NEGATE (0x81) rule sat behind an unreachable branch.
- daemon: blk_submit.c header-sync guard counts fork depth, not header count
  (a node from genesis could never pass 100k headers); download workers get a
  low-block-rate escape for the byte floor; -Wall -Werror across every C rule
  (~393 sites, several real: tx_accept u64 out-params, main.c catch-up slot
  stride 64 vs 80, rpc_server short header write, rpc_wallet_ops descriptor
  truncation, archive_verify ignored prune-mark pwrite).
- New tests the sweep must now pick up: test_interp_core_vectors,
  test_verify_core_vectors (+verify_core_vectors.h), test_csv_disable_flag,
  fuzz_script_diff, fuzz_verify_diff (the last two need a Core oracle build).

The .S re-ports are NOT in this commit; the port still runs the pre-differential
interpreter semantics.

+++ asm/daemon/tx_accept.c
@@ -487,7 +487,7 @@ static long txacc_sigop_cost(void* mp_area, const u8* tx, unsigned long txlen){
     p += 4;
     int segwit = (end - p >= 2 && p[0] == 0x00 && p[1] == 0x01);
     if (segwit) p += 2;
-    unsigned cc; u64 nin = txacc_varint(&p, end, &cc); if (!cc || nin == 0 || nin > 100000) return -1;
+    u64 cc; u64 nin = txacc_varint(&p, end, &cc);   /* u64: txacc_varint stores 8 bytes through this pointer (was `unsigned`: a 4-byte local overwritten by 8 -- 2026-09-02) */ if (!cc || nin == 0 || nin > 100000) return -1;
     struct { const u8* prev; u32 idx; const u8* ss; unsigned long ssl; } in[512];
     if (nin > 512) return -1;                                     /* far above standardness */
     for (u64 i = 0; i < nin; i++){
@@ -559,7 +559,7 @@ long txacc_legacy_sigops(const u8* tx, unsigned long txlen){
     p += 4;
     int segwit = (end - p >= 2 && p[0] == 0x00 && p[1] == 0x01);
     if (segwit) p += 2;
-    unsigned cc; u64 nin = txacc_varint(&p, end, &cc); if (!cc) return -1;
+    u64 cc; u64 nin = txacc_varint(&p, end, &cc);   /* u64: txacc_varint stores 8 bytes through this pointer (was `unsigned`: a 4-byte local overwritten by 8 -- 2026-09-02) */ if (!cc) return -1;
     long total = 0;
     for (u64 i = 0; i < nin; i++){
         if ((unsigned long)(end - p) < 36) return -1;
@@ -587,7 +587,7 @@ const char* txacc_witness_standard(void* mp_area, const u8* tx, unsigned long tx
     int segwit = (end - p >= 2 && p[0] == 0x00 && p[1] == 0x01);
     if (!segwit) return 0;
     p += 2;
-    unsigned cc; u64 nin = txacc_varint(&p, end, &cc); if (!cc || nin == 0 || nin > 512) return 0;
+    u64 cc; u64 nin = txacc_varint(&p, end, &cc);   /* u64: txacc_varint stores 8 bytes through this pointer (was `unsigned`: a 4-byte local overwritten by 8 -- 2026-09-02) */ if (!cc || nin == 0 || nin > 512) return 0;
     struct { const u8* prev; u32 idx; const u8* ss; unsigned long ssl; } in[512];
     for (u64 i = 0; i < nin; i++){
         if ((unsigned long)(end - p) < 36) return 0;
